Data cleanup and some FE stuff

+++ backend/app/models/salary.py
@@ -5,9 +5,9 @@ class ReportedSalary(SQLModel, table=True):
     id: int | None = Field(default=None, primary_key=True)
     company: str
     year: int
-    salary: int
+    salary: float
     university: str 
     term: int | None = None
     location: str
-    bonus: int | None = None
+    bonus: float | None = None
     role: str
